Windows compatibility fixes

+++ classes/pen/pen.cpp
@@ -18,8 +18,9 @@ void Pen::write_on_paper(string words)
 void Pen::break_in_half()
 {
     InkLevelPercent = InkLevelPercent / 2;
-    Length = Length / 2.0;
+    Length = Length / (float)2.0;
 }
+
 void Pen::run_out_of_ink()
 {
     InkLevelPercent = 0;
